+++ src/shared/ui/form/FormField.tsx
@@ -1,14 +1,8 @@
 'use client';
 import { cn } from '@/shared/utils';
-import { ReactNode } from 'react';
+import { FormFieldProps } from './types';
 
-export default function FormField({
-  className,
-  children,
-}: {
-  className?: string;
-  children: ReactNode;
-}) {
+export default function FormField({ className, children }: FormFieldProps) {
   return (
     <div className={cn('mb-6 flex flex-col gap-2', className)}>{children}</div>
   );
