+++ src/shared/ui/form/FormLabel.tsx
@@ -1,12 +1,6 @@
-import React from 'react';
+import { FormLabelProps } from './types';
 
-export default function FormLabel({
-  required,
-  children,
-}: {
-  required?: boolean;
-  children: React.ReactNode;
-}) {
+export default function FormLabel({ required, children }: FormLabelProps) {
   return (
     <label className="text-sm text-gray-700 flex items-center text-body_medium">
       {required && <span className="mr-0.5">*</span>}
