+++ src/shared/ui/form/index.ts
@@ -1,8 +1,9 @@
 import type * as React from 'react';
-import InternalForm, { FormProps } from './Form';
+import InternalForm from './Form';
 import FormContent from './FormContent';
 import FormField from './FormField';
 import FormLabel from './FormLabel';
+import { FormProps } from './types';
 
 type CompoundedComponent = React.FC<FormProps> & {
   Field: typeof FormField;
